+++ chan_lcr.h
@@ -14,7 +14,7 @@ struct bchannel;
 struct chan_call {
 	struct chan_call	*next;	/* link to next call instance */
 	int			state;	/* current call state CHAN_LCR_STATE */
-	unsigned long		ref;	/* callref for this channel */
+	unsigned int		ref;	/* callref for this channel */
 	void			*ast;	/* current asterisk channel */
 	int			pbx_started;
 					/* indicates if pbx que is available */
