+++ endpointapp.cpp
@@ -35,12 +35,12 @@ int EndpointApp::handler(void)
 
 /* mini application for test purpose only */
 
-void EndpointApp::ea_message_port(unsigned long port_id, int message_type, union parameter *param)
+void EndpointApp::ea_message_port(unsigned int port_id, int message_type, union parameter *param)
 {
 	PDEBUG(DEBUG_EPOINT, "%s: Spare function.\n", __FUNCTION__);
 }
 
-void EndpointApp::ea_message_join(unsigned long join_id, int message_type, union parameter *param)
+void EndpointApp::ea_message_join(unsigned int join_id, int message_type, union parameter *param)
 {
 	PDEBUG(DEBUG_EPOINT, "%s: Spare function.\n", __FUNCTION__);
 }
