+++ endpointapp.h
@@ -19,7 +19,7 @@ class EndpointApp
 
 	class Endpoint		*ea_endpoint;
 	virtual int		handler(void);
-	virtual void ea_message_port(unsigned long port_id, int message, union parameter *param);
-	virtual void ea_message_join(unsigned long join_id, int message, union parameter *param);
+	virtual void ea_message_port(unsigned int port_id, int message, union parameter *param);
+	virtual void ea_message_join(unsigned int join_id, int message, union parameter *param);
 };
