+++ join.h
@@ -23,16 +23,16 @@ class Join
 	Join();
 	virtual ~Join();
 	class Join *next;		/* next node in list of joins */
-	virtual void message_epoint(unsigned long epoint_id, int message, union parameter *param);
+	virtual void message_epoint(unsigned int epoint_id, int message, union parameter *param);
 	virtual int handler(void);
 
-	unsigned long j_type;		/* join type (pbx or asterisk) */
-	unsigned long j_serial;		/* serial/unique number of join */
+	unsigned int j_type;		/* join type (pbx or asterisk) */
+	unsigned int j_serial;		/* serial/unique number of join */
 }; 
 
 void join_free(void);
 
 extern class Join *join_first;
 
-class Join *find_join_id(unsigned long join_id);
+class Join *find_join_id(unsigned int join_id);
